Enable LogicalPlanner for simple selects on system tables
